% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{intmap-imports}
\alias{intmap-imports}
\alias{from_just}
\title{Extract value from a 'Just' value}
\description{
The \code{from_just} function is imported from the 
  \strong{maybe} package.
  Follow the link to its documentation: 
  \code{\link[maybe:from_just]{from_just}}. 
  It has been imported for convenient use of the \code{intmap$at} method, 
  which returns a 'Just' value.
}
